Remove final from Channel, and make methods of channel interface to be virtual