@vjpai 2019-07-19
Approval pending approved gRFC. Please consider adding a usage example in the interception documentation.