@yashykt 2019-07-26
The benchmarks don't show any worthwhile changes - 
name                                                                    old time/op             new time/op             delta
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/0               12.7µs ±10%             12.5µs ± 6%   ~             (p=1.000 n=3+3)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/1               4.98µs ± 3%             4.90µs ± 1%   ~             (p=0.700 n=3+3)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/8               4.96µs ± 3%             4.89µs ± 1%   ~             (p=1.000 n=3+3)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/64              5.12µs ± 3%             5.35µs ±12%   ~             (p=1.000 n=3+3)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/512             5.18µs ± 3%             5.12µs ± 2%   ~             (p=1.000 n=3+3)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/0                   1.33µs ± 6%             1.29µs ± 1%   ~             (p=1.000 n=3+3)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/1                   1.35µs ± 6%             1.32µs ± 1%   ~             (p=0.700 n=3+3)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/8                   1.39µs ± 6%             1.35µs ± 1%   ~             (p=0.700 n=3+3)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/64                  1.46µs ± 6%             1.42µs ± 2%   ~             (p=0.400 n=3+3)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/512                 1.51µs ± 5%             1.47µs ± 1%   ~             (p=0.700 n=3+3)

name                                                                    old allocs/op           new allocs/op           delta
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/0                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/1                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/8                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/64                 0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/512                0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/0                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/1                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/8                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/64                     0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/512                    0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)

name                                                                    old peak-mem(Bytes)/op  new peak-mem(Bytes)/op  delta
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/0                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/1                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/8                  0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/64                 0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<InsecureChannelFixture>/512                0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/0                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/1                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/8                      0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/64                     0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
BM_InsecureChannelCreateDestroy<LameChannelFixture>/512                    0.00 ±NaN%              0.00 ±NaN%   ~     (all samples are equal)
The first run seems to be more noisy